split out `TransactionInfo`
