+++ lib/core/cardano-wallet-core.cabal
@@ -278,6 +278,8 @@ library
       Cardano.Wallet.Primitive.Types.Tx
       Cardano.Wallet.Primitive.Types.Tx.CBOR
       Cardano.Wallet.Primitive.Types.Tx.CBOR.Hash
+      Cardano.Wallet.Primitive.Types.Tx.TransactionInfo
+      Cardano.Wallet.Primitive.Types.Tx.Tx
       Cardano.Wallet.Primitive.Types.UTxO
       Cardano.Wallet.Primitive.Types.UTxOIndex
       Cardano.Wallet.Primitive.Types.UTxOIndex.Internal
